% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barra_apilada.R
\name{barra_apilada}
\alias{barra_apilada}
\title{Funcion para realizar un grafico de barra apilada de 1 variable.}
\usage{
barra_apilada(
  data,
  ...,
  filtrar = TRUE,
  ordenado = TRUE,
  porcentaje = TRUE,
  ext.label = 42,
  base.posicion = 0.35
)
}
\arguments{
\item{data}{Base de datos para la funcion}

\item{var}{Variable para el grafico de barra apilada}
}
\value{
Un grafico de barra apilado
}
\description{
Funcion para realizar un grafico de barra apilada de 1 variable.
}
\examples{

# data_prueba_ruta <- system.file("data_prueba.sav", package = "pulso")
# data_prueba <- haven::read_sav(data_prueba_ruta)
# 
# data_prueba \%>\% 
#   filter(q0002 \%in\% 1) \%>\% 
#   barra_apilada_1(q0003_0002)


}
