% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barra_apilada_1.R
\name{barra_apilada_1}
\alias{barra_apilada_1}
\title{Funcion para realizar un grafico de barra apilada de 1 variable.}
\usage{
barra_apilada_1(data, var)
}
\arguments{
\item{data}{Base de datos para la funcion}

\item{var}{Variable para el grafico de barra apilada}
}
\value{
Un grafico de barra apilado
}
\description{
Funcion para realizar un grafico de barra apilada de 1 variable.
}
\examples{


data_prueba \%>\% 
  filter(q0002 \%in\% 1) \%>\% 
  barra_apilada_1(q0003_0002)


}
