% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barra_multiple.R
\name{barra_multiple}
\alias{barra_multiple}
\title{Funcion para realizar un grafico de barras multiple}
\usage{
barra_multiple(
  data,
  ...,
  filtrar = TRUE,
  abierta = FALSE,
  frecuencia = FALSE,
  porcentaje = TRUE,
  ultimo = NULL,
  max.limit = 1,
  color = "#B0D597",
  ext.label = 30
)
}
\arguments{
\item{data}{Base de datos para la funcion}

\item{...}{Variables para el grafico de barras multiple}

\item{filtrar}{Por default TRUE que significa ningún filtro, si se quiere filtrar por una variable especificar la variable y el valor a filtrar (ej: filtrar = q0002 == 1 que significa filtrar la base para que aparezcan solo los casos que tienen en la pregunta q0002 el valor de 1).}

\item{abierta}{Por default abierta es FALSE, es necesario cambiarla a TRUE cuando se trabaja con una pregunta abierta lo cual tendra un impacto en las etiquetas.}

\item{frecuencia}{Por default es FALSE, cambiar a TRUE para visualizar los resultados en frecuencias y no en porcentajes.}

\item{porcentaje}{Por default es TRUE, cambiar a FALSE para visualizar los resultados de porcentajes sin el simbolo \%.}

\item{ultimo}{Por default es NULL, escribir entre comillas la palabra que quieres que vaya al final de las barras (ej: ultimo="Otros")}

\item{max.limit}{Por default es 1, es el limite máximo del eje x. 1=100 cuando son resultados en porcentajes (ej: max.limit=0.75 significa maximo 75\%). Aunque el porcentaje esté en FALSE, el max.limit debe ser tratado como si 1.00 fuera el maximo y no como si 100 fuera el maximo.}

\item{color}{Por default es color azul que es "#B0D597" en notacion hexagesimal. Buscar colores hex para más información. También permite colores grabados en R como "red".}

\item{ext.label}{Por default es 30, a menor el número menor espacio para el texto de las etiquetas.}
}
\value{
Un grafico de barras múltiple
}
\description{
Funcion para realizar un grafico de barras multiple
}
\examples{


# data_prueba \%>\% 
#   filter(q0002 \%in\% 1) \%>\% 
#   barra_multiple(starts_with("q0004_0"))


}
