% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_donut.R
\name{grafico_donut}
\alias{grafico_donut}
\title{Funcion para realizar un grafico donut.}
\usage{
grafico_donut(data, var, paleta = 1, direction = 1)
}
\arguments{
\item{data}{Base de datos para la funcion}

\item{var}{Variable para gráfico donut}
}
\value{
Un grafico donut
}
\description{
Funcion para realizar un grafico donut.
}
\examples{

data_prueba \%>\% 
  grafico_donut(sexo)

}
