% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_pie.R
\name{grafico_pie}
\alias{grafico_pie}
\title{Funcion para realizar un grafico pie}
\usage{
grafico_pie(data, var, filtrar = TRUE, color1 = "#A9D18E", color2 = "#FF9393")
}
\arguments{
\item{data}{Base de datos para la funcion}

\item{var}{Variable para gráfico pie}
}
\value{
Un grafico pie
}
\description{
Funcion para realizar un grafico pie
}
\examples{

# data_prueba_ruta <- system.file("data_prueba.sav", package = "pulso")
# data_prueba <- haven::read_sav(data_prueba_ruta)
# 
# data_prueba \%>\% 
#   grafico_donut(sexo)

}
