% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_radar.R
\name{grafico_radar}
\alias{grafico_radar}
\title{Funcion para realizar un grafico radar}
\usage{
grafico_radar(
  plot.data,
  base.size = 20,
  webtype = "mini",
  axis.labels = "",
  grid.min = 0,
  grid.max = 1,
  centre.y = grid.min - ((1/9) * (grid.max - grid.min)),
  label.centre.y = FALSE,
  grid.line.width = 0.5,
  grid.line.trend = "classic",
  gridline.min.linetype = "longdash",
  gridline.mid.linetype = "longdash",
  gridline.max.linetype = "longdash",
  gridline.min.colour = "grey",
  gridline.mid.colour = "#007A87",
  gridline.max.colour = "grey",
  grid.label.size = 6,
  gridline.label.offset = -0.1 * (grid.max - centre.y),
  label.gridline.min = TRUE,
  label.gridline.mid = TRUE,
  label.gridline.max = TRUE,
  gridline.label = NULL,
  axis.label.offset = 1.15,
  axis.label.size = 5,
  axis.line.colour = "grey",
  group.line.width = 1.5,
  group.point.size = 3,
  group.colours = NULL,
  group.fill.colours = NULL,
  background.circle.colour = "#D7D6D1",
  background.circle.transparency = 0.2,
  legend.title = "",
  plot.legend = TRUE,
  plot.title = "",
  legend.text.size = 14,
  radarshape = "round",
  polygonfill = TRUE,
  polygonfill.transparency = 0.2,
  multiplots = FALSE,
  stripbackground = TRUE,
  fullscore = NULL
)
}
\arguments{
\item{data}{Base de datos para la funcion}
}
\value{
Un grafico radar
}
\description{
Funcion para realizar un grafico radar
}
\examples{
\dontrun{
radar \%>\% 
  select(group, Becas, `Atención socio- económica`, Empleabilidad, `Tutorías`) \%>\% 
  grafico_radar(polygonfill = FALSE,
                grid.label.size = 3,
                axis.label.size = 3,
                group.line.width = 1,
                fullscore = as.numeric(rep(1,ncol(.)-1)),
         ) + 
  
  labs(caption = "Elaborado por Pulso PUCP",
       tag = glue("N=",radar.tag)) +
  
  theme(legend.title = element_blank(),legend.position = "bottom",legend.text = element_text(size=10, face = "bold",family="sans"),legend.key.height = unit(.2, "cm"),
        plot.caption = element_text(face = "italic",family="sans"),plot.margin = unit(c(0,0,1,0),"cm"),plot.tag = element_text(size = 8, color="grey40"),plot.tag.position = "topright",
        
        text = element_text(size = 9, color="#002060",family="sans"),
        ) +
  guides(color=guide_legend(nrow = 2, byrow = TRUE)) +
  coord_equal(clip="off")
}
}
