% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulso-package.R
\docType{package}
\name{pulso-package}
\alias{pulso}
\alias{pulso-package}
\title{pulso: Load Microdata from Colombia's GEIH (DANE)}
\description{
R companion to the pulso-co Python package. Provides programmatic
access to microdata from Colombia's Gran Encuesta Integrada de Hogares
(GEIH), published by DANE.
}
\section{Main functions}{

\itemize{
\item \code{\link[=pulso_load]{pulso_load()}} - Load microdata for a single year/month/module
\item (More to come in v0.1.0+)
}
}

\section{Comparison with Python}{

This package mirrors the API of the Python \code{pulso-co} package. See
the vignette for details.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Stebandido77/pulso}
  \item Report bugs at \url{https://github.com/Stebandido77/pulso/issues}
}

}
\author{
\strong{Maintainer}: Esteban Labastidas \email{estebanlabastidas123@gmail.com}

Authors:
\itemize{
  \item Esteban Labastidas \email{estebanlabastidas123@gmail.com}
}

}
\keyword{internal}
