% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barra_apilada_n.R
\name{barra_apilada_n}
\alias{barra_apilada_n}
\title{Funcion para realizar un grafico de barra apilada de n variables.}
\usage{
barra_apilada_n(data, ..., ordenado = TRUE)
}
\arguments{
\item{data}{Base de datos para la funcion}

\item{...}{Variables para el grafico de barra apilada}
}
\value{
Un grafico de barras apiladas
}
\description{
Funcion para realizar un grafico de barra apilada de n variables.
}
\examples{
base \%>\% 
  barra_apilada_n(?starts_with("q0011_"))
}
