% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barra_multiple.R
\name{barra_multiple}
\alias{barra_multiple}
\title{Funcion para realizar un grafico de barras multiple}
\usage{
barra_multiple(
  data,
  ...,
  subtitulo = "",
  abierta = nombres,
  max.limit = 1,
  ultimo = NULL,
  color = "#B0D597",
  ext.label = 30
)
}
\arguments{
\item{data}{Base de datos para la funcion}

\item{...}{Variables para el grafico de barras multiple}
}
\value{
Un grafico de barras múltiple
}
\description{
Funcion para realizar un grafico de barras multiple
}
\examples{


data_prueba \%>\% 
  filter(q0002 \%in\% 1) \%>\% 
  barra_multiple(starts_with("q0004_0"))


}
