% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barra_simple.R
\name{barra_simple}
\alias{barra_simple}
\title{Funcion para realizar un grafico de barras simple.}
\usage{
barra_simple(
  data,
  var,
  ordenado = TRUE,
  max.limit = 1,
  color = "#B0D597",
  ultimo = NULL
)
}
\arguments{
\item{data}{Base de datos para la funcion}

\item{var}{Variable para el grafico de barras simple}
}
\value{
Un grafico de barras simple.
}
\description{
Funcion para realizar un grafico de barras simple.
}
\examples{
\dontrun{
base \%>\% 
  barra_simple(q0010)

barra_simple(base, q0010)
}
}
