% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulso-package.R
\docType{package}
\name{pulso-package}
\alias{pulso}
\alias{pulso-package}
\title{pulso: Funciones Para Los Informes Pulso PUCP}
\description{
Conjunto de funciones para los informes de Pulso PUCP.
}
\author{
\strong{Maintainer}: Santiago Sotelo \email{santiago.sotelo@pucp.edu.pe} (\href{https://orcid.org/0000-0002-9739-9964}{ORCID})

}
\keyword{internal}
